+++ lib/src/cli/commands/extract_command.dart
@@ -80,8 +80,15 @@ If not specified the extracted API will be printed to the console.
       doRemoveExample: doRemoveExample,
     );
     await cleanUp(preparedPackageRef);
-    final jsonString =
-        PackageApiStorage.packageApitoStorageJson(packageApi, pretty: true);
+
+    final declarationsWithoutEntryPoints =
+        packageApi.rootDeclarationsWithoutEntryPoints;
+
+    final jsonString = PackageApiStorage.packageApitoStorageJson(
+      packageApi,
+      declarationsWithoutEntryPoints,
+      pretty: true,
+    );
     final outFilePath = argResults![_optionNameOutput] as String?;
     if (outFilePath != null) {
       final outFile = File(outFilePath);
@@ -94,13 +101,13 @@ If not specified the extracted API will be printed to the console.
       stdout.writeln(jsonString);
     }
 
-    final declarationsWithoutEntryPoints =
-        packageApi.getRootDeclarationsWithoutEntryPoints();
     if (declarationsWithoutEntryPoints.isNotEmpty) {
-      stdout.writeln(
-          'The following declarations do not have an entry point (did you miss to export them?):');
-      for (final declaration in declarationsWithoutEntryPoints) {
-        stdout.writeln('  ${declaration.name}');
+      if (outFilePath == null) {
+        stdout.writeln(
+            'The following declarations do not have an entry point (did you miss to export them?):');
+        for (final declaration in declarationsWithoutEntryPoints) {
+          stdout.writeln('  ${declaration.name}');
+        }
       }
       if (doSetExitCodeOnMissingExport) {
         return -1;