+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - bump dart SDK requirements to >=3.0.0
 - Search downloaded packages in entire .pub-cache hosted directory in case path namining changes
 - adds information about the package versions and the version check result to the reports (console, json, markdown)
+- add missing export to json output for `extract` command
 
 ## Version 0.16.3
 - adds more diff result reporting options (cli, json, markdown)